requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
requests
werkzeug
google-generativeai
typing-extensions
python-dotenv